devices/cameras/person-detection/requirements.txt
appdirs==1.4.4
backports.entry-points-selectable==1.1.0
black==20.8b1
boto3==1.18.44
botocore==1.21.44
certifi==2021.5.30
cfgv==3.3.1
charset-normalizer==2.0.6
click==8.0.1
colorlog==4.8.0
dacite==1.6.0
distlib==0.3.2
filelock==3.0.12
flake8==3.9.2
gitdb==4.0.7
GitPython==3.1.24
identify==2.2.15
idna==3.2
isort==5.9.3
jmespath==0.10.0
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.21.2
opencv-python==4.5.3.56
paho-mqtt==1.5.1
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.3.0
pre-commit==2.15.0
pycodestyle==2.7.0
pyflakes==2.3.1
pysen==0.9.1
python-dateutil==2.8.2
python-dotenv==0.19.0
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
smmap==4.0.0
toml==0.10.2
tomli==1.2.1
tomlkit==0.7.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
unidiff==0.7.0
urllib3==1.26.6
virtualenv==20.8.0
